+++ src/pages/ChatsPage.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {Grid, Typography} from "@mui/material";
-import ChatList from "../components/ChatList";
+import ChatListContainer from "../components/ChatListContainer";
 import Messages from "../components/Messages";
 //import {useParams} from "react-router-dom";
 
@@ -13,7 +13,7 @@ const ChatsPage = () => {
 
                     <Grid item xs={4} md={2}>
                         <Typography variant="h5" component="div" color="primary">Chat list</Typography>
-                        <ChatList/>
+                        <ChatListContainer/>
                     </Grid>
 
                     <Grid item xs={8} md={6}>
